backend/pyproject.toml
[project]
name = "crewboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.132.0",
    "httpx>=0.28.1",
    "psycopg[binary]>=3.3.3",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "rich>=14.3.3",
    "sqlalchemy>=2.0.46",
    "tenacity>=9.1.4",
    "uvicorn[standard]>=0.41.0",
]
